requirements.txt
absl-py==0.13.0
aenum==2.2.0
astor==0.8.1
attrs==19.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
decorator==4.4.0
descartes==1.1.0
enum34==1.1.6
Fiona==1.8.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.2.2
geographiclib==1.49
geopandas==0.5.1
geopy==1.20.0
google-auth==1.32.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpxpy==1.3.5
grpcio==1.38.1
h5py==2.10.0
idna==2.8
ipython==5.8.0
ipython-genutils==0.2.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mock==3.0.5
mplleaflet==0.0.5
munch==2.3.2
networkx==2.2
numpy==1.16.4
oauthlib==3.1.0
opt-einsum==2.3.2
osmnx==0.9
pandas==0.24.2
pathlib2==2.3.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==1.0.16
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.4.2
pyparsing==2.4.7
pyproj==2.2.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.5
Rtree==0.8.3
scandir==1.10.0
scipy==1.2.2
Shapely==1.6.4.post2
simplegeneric==0.8.1
simpy==3.0.11
six==1.12.0
smopy==0.0.7
subprocess32==3.5.4
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==1.0.1
wrapt==1.12.1
